requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pulp>=2.9.0
ortools>=9.10.0
streamlit>=1.32.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-pptx>=1.0.2
